refactor(MessageBar): move motions to own file
